rename: unify `typed` and `format` module
