+++ ohkami/src/typed/body.rs
@@ -1,0 +1,14 @@
+mod json;
+pub use json::Json;
+
+mod multipart;
+pub use multipart::{Multipart, File};
+
+mod urlencoded;
+pub use urlencoded::UrlEncoded;
+
+mod text;
+pub use text::Text;
+
+mod html;
+pub use html::Html;
